requirements.txt
plotly
pandas
numpy
scikit-learn
statsmodels
pyyaml
xgboost
streamlit
mlflow
